Tweak install deps script to allow installers to be run in series as …

+++ install_deps.sh
@@ -2,5 +2,17 @@
 
 set -e
 
-docker-compose -f docker-compose.yarn.yml up
-docker-compose -f docker-compose.yarn.yml down
+if [[ "$1" == "--series" ]];
+then
+  containers=("zen" "frontend" "badges" "dojos" "events"
+    "eventbrite" "organisations" "users" "events_service"
+    "users_service" "clubs_service" "email_service")
+  
+  for container in ${containers[@]};
+  do
+    docker-compose -f docker-compose.yarn.yml run yarn_${container}
+  done
+else
+  docker-compose -f docker-compose.yarn.yml up
+  docker-compose -f docker-compose.yarn.yml down
+fi
